requirements.txt
python-jobspy>=0.1.0
requests>=2.31.0
streamlit>=1.35.0
